Confirmar exclusão de imagem

+++ funcoes/apagarImagem.php
@@ -1,10 +1,15 @@
 <?php 
 
     $imagem = $_GET['imagem'];
-    $nome_banner = $_GET['banner'];
-    unlink($imagem);
-    header('Location: ../view/uploadImagem.php?banner='.$nome_banner.'') 
-
-
-
+    if (file_exists($imagem)) 
+    {
+    	unlink($imagem);
+    	echo 1;
+    }
+    else
+    {
+    	echo 2;
+    }
+    
+    
 ?>
